move ConnectionHandlerCellMarker to external file from ConnectionHandler to allow use in touch story example
